+++ ppdet/data/source/category.py
@@ -26,7 +26,7 @@ logger = setup_logger(__name__)
 __all__ = ['get_categories']
 
 
-def get_categories(metric_type, arch, anno_file=None):
+def get_categories(metric_type, anno_file=None, arch=None):
     """
     Get class id to category id map and category id
     to category name map from annotation file.
